Separates out Map creator helper functions
