+++ src/components/Map/Map.js
@@ -1,16 +1,8 @@
 import React from "react";
+import { MapCreator } from './helpers';
+
 // rendered by Game
-const Map = () => {
-  return (
-    <div>
-      <h1>Map</h1>
-    </div>
-  );
-};
-
-export default Map;
-
-let MUD_Map = MapCreator(5, 10, 64, [[
+const Map = MapCreator(5, 10, 64, [[
   2, 2, 2, 1, 2, 1, 1, 1, 2, 2,
   1, 2, 1, 2, 1, 2, 2, 1, 1, 1,
   2, 1, 2, 1, 1, 2, 1, 2, 2, 2,
@@ -28,34 +20,14 @@ let MUD_Map = MapCreator(5, 10, 64, [[
   11, 0, 0, 0, 0, 20, 0, 12, 0, 19,
   9, 14, 0, 18, 0, 9, 0, 0, 10, 6,
   12, 7, 13, 0, 7, 10, 7, 6, 14, 0
-]])
+]]);
 
-class MapCreator {
-  constructor(rows, columns, tile_size, layers){
-    this.rows = rows;
-    this.columns = columns;
-    this.tile_size = tile_size;
-    this.layers = layers;
-  }
-
-  getTile = (layer, column, row) => {
-    return this.layers[layer][row * this.columns + column];
-  };
-
-  getColumn = (x) => {
-    return Math.floor(x / this.tile_size);
-  };
-
-  getRow = (y) => {
-    return Math.floor(y / this.tile_size);
-  };
-
-  getX = (column) => {
-    return column * this.tile_size;
-  };
-
-  getY = (row) => {
-    return row * this.tile_size;
-  };
-
+const MapComponent = () => {
+  return (
+    <div>
+      <h1>Map</h1>
+    </div>
+  );
 };
+
+export default Map;
